requirements.txt
mlflow>=2.8.0
dagshub>=0.3.25
pandas>=1.3.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.21.0
python-dotenv>=1.0.0
requests>=2.28.0